Revert "fix(planning): clip capacity bar fill at cap, mark overflow inside bar"
